+++ app/models/project.rb
@@ -230,25 +230,6 @@ class Project < ApplicationRecord
     AUTO_LOCK_SHEETS.find { |_name, value| value == auto_lock_sheets }.first
   end
 
-  def runnable_checks
-    checks.where(archived: false).where.not(message: [nil, ''])
-  end
-
-  def failed_sheet_id_checks(current_user, filtered_sheets)
-    hash = {}
-    filtered_sheets.each do |sheet|
-      hash[sheet.id] = []
-    end
-    runnable_checks.find_each do |check|
-      sheet_ids = check.sheets(current_user).pluck(:id)
-      sheet_ids.each do |sheet_id|
-        next unless hash.key?(sheet_id)
-        hash[sheet_id] << check
-      end
-    end
-    hash
-  end
-
   private
 
   # Creates a default site if the project has no site associated with it